requirements.txt
SQLAlchemy==1.3.3
pandas==0.24.2
nltk==3.4.1
numpy==1.16.3
matplotlib==3.0.3
tqdm==4.31.1
missingno==0.4.1
boto3==1.9.146
PyYAML==5.1
